requirements.txt
pytest == 8.0.0
allure-pytest == 2.13.2
selene>=2.0.b4
python-dotenv~=1.0.1
selenium~=4.20.0